streamlining data entry
We addressed a glitch with the multispecimen checkbox and found that enter needs to be hit for the value to be registered.
We also made sure that on each selection of the listboxes the focus is automatically skipped on the next field in all cases. When saving, focus is redirected to the barcode field for the next entry.