Fixed container data inconsistencies | v1.1.6
Issue: "6.	When I go back and change from MSO or MOS to single specimen all is fine. However, when I move to the next record in the sequence which had the MSO or MOS it now has the container name in it, but it is “single specimen object” which is selected in the UI."