Fix for #400 "Needs repair" not cleared  | v1.1.11
The way the data entry form is set up for a new record during saving was deviating from the MVC pattern: It no longer instantiated a new instance of the specimen model, but re-used the current instance, yet failing to clear all fields. It's been brought back to the original method of creating a new instance and transferring sticky field values to that instance.