FIx for #407 : Repeated taxon number | v1.1.14
There was indeed a problem with the source file of the taxon numbers. Those are now fixed and the SQL is rerun for the Specify database and then the SQL for the app db was regenerated on the basis of that, setting the correct taxon numbers. One issue was the auxiliary taxon number having been removed e.g. "1166,2" had become "1166"...