#407 Fixed binomial names with missing genera
There was a limited number of cases where the genus name was not added to the full name, which lead to the SQL script (MassDigitizer/sql/specify/UpdateSetTaxonNumber.sql) failing to add those taxon numbers to the taxa in question, since it relied on full name (binomial). The latter file is now fixed by having redone the binomial full name in those cases.