Hack for #412 preventing infinite loop | v1.1.17
Shift-tabbing with activated taxon number field caused the app to go in an infinite loop constantly switching between setting focus on taxon number field and barcode field. The cause of this bug could not be traced, but the infinite loop is now stopped in its tracks by measuring the amount of time elapsed and breaking when it is shorter that 50 ms.

Another major change that was made with the idea to prevent UI triggers that would the possible cause is getting the form values from events instead of directly from the window. It may need to be tested, however, if all values are still correctly persisted.