Fix for #420 Persisting changed container name | v1.1.18
The container "ID" or name was not updated when editing an older record. Event handler was added and made to be triggered whenever the field is edited or even just left again. Now when pressing save, the new value is persisted to the record.

The field name was also redubbed "inpContainerName" to better reflect the Specify concept.