Taxon Spine for VASC as to help #434
The Taxon Spine is included as tsv files that will be the eventual source for the taxon names that the digi app will be able to suggest to the end user. What is lacking still is some way to read in these name records and insert these into the local app db.