Fix for #466 : Can't add new taxa without author | v1.1.25
This fix enables the escape key and will then close the popup while transferring whatever has already been written to the taxon input field in the main data entry form. At the same time it will remove the character "*". This way, when there is a match with author name that the end user does not want to include, they can simple add the asterisk and then escape.