Fix for #474 : Crash when entering taxon number | v1.1.27
Simple fix for the NHMA taxon number crash. Code attempted to create a new taxon instances and setting the corresponding fields in a generic model. The generic model for some reason expected a field with name 'collectionid', which does not exist in the table in question (taxonname). This is also unncessary to set, since the model is always initiated with a collectionid first anyway. The line in question is therefore outcommented.