Fix to post-processing script: New taxon handling
The previous changes to the post-processing script centered on the dasscoid in order to determine new taxa. However, not all taxon spines include the dasscoid (yet), so we still need to default on the taxonspid. This fix checks on both dasscoid and taxonspid to assess whether a taxon is known.