Added lacking filtering on taxon tree and institution in Full Text Search #441
Virtual FTS (Full Text Search) table meant to facilitate full text search did not include columns for institution and taxontree id, so the search was not limited to the current collection, but crossed all collections. With this fix these columns were added to the virtual table denoting institution and taxontree and corresponding filters were added to the FTS search.