Fix for #479 | Tested & tweaked GREL script for NHMA imports
Taxon numbers were already ensured in a previously committed fix. I just pruned some parts of that, since infraspecific taxa are not assigned taxon number keys in NHMA.

NHMA also turned out to have a different format for dates, which affects workbench import of the cataloged date. So that has been adapted for that institution only, by filtering on institutionid.