Major overhaul of UI for #439, #445, #274
New features:
* Rearranged data input fields to a new flow of direction (#439)
* Added tick boxes for specimen obscured and label obscured (#445)
* Added multi-object warning whenever a linked record is created or viewed (#274)
* Enable rank determination of taxon name also when author is added (#484)

The latter feature (#484) requires the end user to separate out the authorship section of a novel taxon name entry using an underscore and an underscore only e.g. "Eriocrana incognita_Schütler 1984". This way the authorship can be taken out before parsing the taxon name in order to determine rank. Rank determination is dependent on a set pattern of space-separated elements of the full binomial or trinomial name.